Blocking and Parry advances
When player blocks is invulnerabel for now, we need to create a resistance bar to not abuse this mechanic. If block at the right time(first part of teh anim) the player does a parry that "stuns" enemy for 1 sec aprox. BUGS: Player can block saw and blade traps if he blocks. Could leave it and test how it feels with the blocking bar(sekiro)